requirements.txt
certifi==2020.6.20
Django==2.2.15
django-filter==2.3.0
djangorestframework==3.11.0
importlib-metadata==1.7.0
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0

requests~=2.24.0
jsonpath~=0.82